Add fee information in mempool's transaction accumulator
